ch(ID handler): added ID handler middle-ware - added the id handler file - added the params middle-ware for entry id - made test for non-numeric IDs [finishes #169227826]
