--- .nyc_output/processinfo/9973ab4c-f159-42d0-813b-2f334402181c.json
@@ -1,1 +1,0 @@
-{"uuid":"9973ab4c-f159-42d0-813b-2f334402181c","parent":null,"pid":5188,"argv":["C:\\Program Files\\nodejs\\node.exe","C:\\Users\\fina\\Desktop\\mydiary-website\\node_modules\\mocha\\bin\\mocha","--require","@babel/register","--require","@babel/polyfill","--timeout","100000","Server/tests/","--exit"],"execArgv":[],"cwd":"C:\\Users\\fina\\Desktop\\mydiary-website","time":1571247085226,"ppid":6308,"root":"1fa2eb69-d79c-4a2d-8207-f3904e799b9a","coverageFilename":"C:\\Users\\fina\\Desktop\\mydiary-website\\.nyc_output\\9973ab4c-f159-42d0-813b-2f334402181c.json","files":["C:\\Users\\fina\\Desktop\\mydiary-website\\Server\\app.js","C:\\Users\\fina\\Desktop\\mydiary-website\\Server\\routes\\userRoute.js","C:\\Users\\fina\\Desktop\\mydiary-website\\Server\\controllers\\userController.js","C:\\Users\\fina\\Desktop\\mydiary-website\\Server\\helpers\\send.js","C:\\Users\\fina\\Desktop\\mydiary-website\\Server\\data\\userData.js","C:\\Users\\fina\\Desktop\\mydiary-website\\Server\\helpers\\userValidators.js","C:\\Users\\fina\\Desktop\\mydiary-website\\Server\\routes\\entryRoute.js","C:\\Users\\fina\\Desktop\\mydiary-website\\Server\\controllers\\entryController.js","C:\\Users\\fina\\Desktop\\mydiary-website\\Server\\data\\entryData.js","C:\\Users\\fina\\Desktop\\mydiary-website\\Server\\helpers\\entryValidators.js","C:\\Users\\fina\\Desktop\\mydiary-website\\Server\\helpers\\protect.js"]}
