+++ .nyc_output/processinfo/e48116ed-fba4-4f53-87e2-3178899cfebc.json
@@ -1,0 +1,1 @@
+{"uuid":"e48116ed-fba4-4f53-87e2-3178899cfebc","parent":null,"pid":8132,"argv":["C:\\Program Files\\nodejs\\node.exe","C:\\Users\\fina\\Desktop\\mydiary-website\\node_modules\\mocha\\bin\\mocha","--require","@babel/register","--require","@babel/polyfill","--timeout","100000","Server/tests/","--exit"],"execArgv":[],"cwd":"C:\\Users\\fina\\Desktop\\mydiary-website","time":1571398515641,"ppid":6732,"root":"480298e5-588a-4da9-bd29-34c4ef6358dc","coverageFilename":"C:\\Users\\fina\\Desktop\\mydiary-website\\.nyc_output\\e48116ed-fba4-4f53-87e2-3178899cfebc.json","files":["C:\\Users\\fina\\Desktop\\mydiary-website\\Server\\app.js","C:\\Users\\fina\\Desktop\\mydiary-website\\Server\\routes\\userRoute.js","C:\\Users\\fina\\Desktop\\mydiary-website\\Server\\controllers\\userController.js","C:\\Users\\fina\\Desktop\\mydiary-website\\Server\\helpers\\send.js","C:\\Users\\fina\\Desktop\\mydiary-website\\Server\\data\\userData.js","C:\\Users\\fina\\Desktop\\mydiary-website\\Server\\helpers\\userValidators.js","C:\\Users\\fina\\Desktop\\mydiary-website\\Server\\routes\\entryRoute.js","C:\\Users\\fina\\Desktop\\mydiary-website\\Server\\controllers\\entryController.js","C:\\Users\\fina\\Desktop\\mydiary-website\\Server\\data\\entryData.js","C:\\Users\\fina\\Desktop\\mydiary-website\\Server\\helpers\\entryValidators.js","C:\\Users\\fina\\Desktop\\mydiary-website\\Server\\middleware\\protect.js","C:\\Users\\fina\\Desktop\\mydiary-website\\Server\\middleware\\idHandler.js"]}
