+++ Server/routes/entryRoute.js
@@ -1,7 +1,8 @@
 import express from 'express';
 import EntryController from '../controllers/entryController';
 import users from '../data/userData';
-import protect from '../helpers/protect';
+import protect from '../middleware/protect';
+import idhandler from '../middleware/idHandler';
 
 const router = express.Router();
 router.use(protect);
@@ -9,6 +10,7 @@ router
   .route('/')
   .post(EntryController.createEntry)
   .get(EntryController.getAllEntry);
+router.param('id', idhandler);
 router
   .route('/:id')
   .patch(EntryController.updateEntry)
