+++ package.json
@@ -13,15 +13,15 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/izzett222/MyDiary.git"
+    "url": "git+https://github.com/izzett222/mydiary-website.git"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "bugs": {
-    "url": "https://github.com/izzett222/MyDiary/issues"
+    "url": "https://github.com/izzett222/mydiary-website/issues"
   },
-  "homepage": "https://github.com/izzett222/MyDiary#readme",
+  "homepage": "https://github.com/izzett222/mydiary-website#readme",
   "dependencies": {
     "@hapi/joi": "^16.1.7",
     "bcrypt": "^3.0.6",
